+++ lib/components/column/pipen_column.dart
@@ -7,10 +7,11 @@ class PipenColumn extends StatelessWidget {
     this.padding,
     this.vertical,
     this.children,
+    this.listable,
     this.horizontal,
     this.mainAxisSize,
   }) {
-    assert(child != null || children != null);
+    assert(child != null || children != null || listable != null);
   }
 
   factory PipenColumn.center({
@@ -19,32 +20,55 @@ class PipenColumn extends StatelessWidget {
     EdgeInsets? padding,
     List<Widget>? children,
     MainAxisSize? mainAxisSize,
+    Iterable<Widget>? listable,
   }) =>
       PipenColumn(
         key: key,
         child: child,
         padding: padding,
+        listable: listable,
         children: children,
         mainAxisSize: mainAxisSize,
         vertical: MainAxisAlignment.center,
         horizontal: CrossAxisAlignment.center,
       );
 
-  factory PipenColumn.end({Key? key, Widget? child, List<Widget>? children, EdgeInsets? padding}) => PipenColumn(
+  factory PipenColumn.end({
+    Key? key,
+    Widget? child,
+    EdgeInsets? padding,
+    List<Widget>? children,
+    MainAxisSize? mainAxisSize,
+    Iterable<Widget>? listable,
+  }) =>
+      PipenColumn(
         key: key,
         child: child,
         padding: padding,
         children: children,
+        listable: listable,
+        mainAxisSize: mainAxisSize,
         vertical: MainAxisAlignment.end,
         horizontal: CrossAxisAlignment.end,
       );
 
   final CrossAxisAlignment? horizontal;
   final MainAxisAlignment? vertical;
+  final Iterable<Widget>? listable;
   final MainAxisSize? mainAxisSize;
   final List<Widget>? children;
   final EdgeInsets? padding;
   final Widget? child;
+
+  List<Widget> get _children {
+    if (children case List<Widget> children) {
+      return children;
+    } else if (child case Widget child) {
+      return [child];
+    } else {
+      return (listable as Iterable<Widget>).toList();
+    }
+  }
 
   @override
   Widget build(BuildContext context) => Padding(
@@ -53,7 +77,7 @@ class PipenColumn extends StatelessWidget {
           mainAxisSize: mainAxisSize ?? MainAxisSize.max,
           mainAxisAlignment: vertical ?? MainAxisAlignment.start,
           crossAxisAlignment: horizontal ?? CrossAxisAlignment.start,
-          children: children ?? [child!],
+          children: _children,
         ),
       );
 }